Update MainGame_Renderer.cs
The EndTurn Listener should not directly call GameController, as there are potentially several render actions that also need to be completed on turn end (such as turning of the glowing ring around the end turn button). The Listener was modified to call EndTurnListener which among other things also calls the GameController.CommitTurn().